[Tm-5] - Calendar style improvements.
