+++ src/src_V2/modules/manager/tour-management/tour-packages/add/calendar/dialogs/GuideDetails.dialog.jsx
@@ -1,7 +1,7 @@
 import { Person } from '@mui/icons-material';
 
-import AddSupplierDetails from './AddSupplierDetails.dialog';
-import { AccommodationSupplier } from './component';
+import AddSupplierDetails from './AddSupplier.dialog';
+import AccommodationSupplier from './accommodation/AccommodationItem';
 import { Accommodation as AccommodationMock } from './mock';
 
 function AddAccommodation(props) {
