+++ src/src_V2/modules/manager/tour-management/tour-packages/add/calendar/dialogs/accommodation/AccommodationDetails.dialog.jsx
@@ -1,0 +1,15 @@
+import { Hotel } from '@mui/icons-material';
+
+import AddSupplierDetails from '../AddSupplier.dialog';
+import AddAccommodation from './Add';
+
+export default function AccommodationDetailsDialog(props) {
+  return (
+    <AddSupplierDetails
+      {...props}
+      title={'Accommodation'}
+      icon={<Hotel color='error' />}
+      SupplierComponent={AddAccommodation}
+    />
+  );
+}
