+++ src/src_V2/modules/manager/tour-management/tour-packages/add/calendar/dialogs/accommodation/Add.jsx
@@ -1,0 +1,17 @@
+import { Accommodation as AccommodationMock } from '../mock';
+import AccommodationItem from './AccommodationItem';
+
+export default function AddAccommodation(props) {
+  return (
+    <>
+      {AccommodationMock.data.map((item, index) => (
+        <AccommodationItem
+          key={index}
+          data={item}
+          expanded={props.expandedId}
+          expandCollapse={props.expandCollapse}
+        />
+      ))}
+    </>
+  );
+}
